1、增加分布式ID 2、增加服务退出监听

+++ models/account.go
@@ -11,8 +11,8 @@ const (
 )
 
 type Account struct {
-	ID         uint        `json:"id" gorm:"column:id"`
-	UserID     uint        `json:"user_id" gorm:"column:user_id"` // 外键，关联 User
+	ID         uint64      `json:"id" gorm:"column:id"`
+	UserID     uint64      `json:"user_id" gorm:"column:user_id"` // 外键，关联 User
 	Identifier string      `json:"identifier"`                    // 可以是手机号、邮箱或其他用户名
 	Password   string      `json:"password"`
 	Type       AccountType `json:"type"`
